Address CR, add debounce in case of unstable connection
