+++ app/components/network_indicator/index.js
@@ -9,7 +9,7 @@ import {init as initWebSocket, close as closeWebSocket} from '@actions/websocket
 import {getCurrentChannelId} from '@mm-redux/selectors/entities/channels';
 
 import {connection} from 'app/actions/device';
-import {markChannelViewedAndReadOnReconnect, setChannelRetryFailed} from 'app/actions/views/channel';
+import {markChannelViewedAndReadOnReconnect, setChannelRetryFailed, loadPostsIfNecessaryWithRetry} from 'app/actions/views/channel';
 import {setCurrentUserStatusOffline, logout} from 'app/actions/views/user';
 import {getConnection, isLandscape} from 'app/selectors/device';
 
@@ -40,6 +40,7 @@ function mapDispatchToProps(dispatch) {
             setCurrentUserStatusOffline,
             startPeriodicStatusUpdates,
             stopPeriodicStatusUpdates,
+            loadPostsIfNecessaryWithRetry,
         }, dispatch),
     };
 }
