+++ app/components/post_list/index.js
@@ -9,7 +9,7 @@ import {getConfig, getCurrentUrl} from '@mm-redux/selectors/entities/general';
 import {getTheme} from '@mm-redux/selectors/entities/preferences';
 import {makePreparePostIdsForPostList, START_OF_NEW_MESSAGES} from '@mm-redux/utils/post_list';
 
-import {handleSelectChannelByName, refreshChannelWithRetry, loadPostsIfNecessaryWithRetry} from 'app/actions/views/channel';
+import {handleSelectChannelByName, refreshChannelWithRetry} from 'app/actions/views/channel';
 import {setDeepLinkURL} from 'app/actions/views/root';
 import {getConnection} from 'app/selectors/device';
 
@@ -44,7 +44,6 @@ function mapDispatchToProps(dispatch) {
             refreshChannelWithRetry,
             setDeepLinkURL,
             showPermalink,
-            loadPostsIfNecessaryWithRetry,
         }, dispatch),
     };
 }
